refact: moving from services to chatAgent
