+++ backend/src/__tests__/chat.test.ts
@@ -24,10 +24,30 @@ import * as chatAgent from '../agents/chatAgent/chatAgent'
 import { ChatMessage } from '@/entities/ChatMessage'
 import { Chat } from '@/entities/Chat'
 import { randomUUID } from 'crypto'
+import { rewrite } from '../agents/chatAgent/queryRewrite'
+import { LLMService } from '../services/llmService'
 
 // Mock external dependencies
 jest.mock('../agents/chatAgent/preprocessing')
 jest.mock('../agents/chatAgent/chatAgent')
+jest.mock('../services/llmService')
+jest.mock('@sentry/node', () => ({
+    captureException: jest.fn(),
+}))
+jest.mock('../lib/logger', () => ({
+    logger: {
+        info: jest.fn(),
+        warn: jest.fn(),
+        error: jest.fn(),
+    },
+}))
+jest.mock('../settings', () => {
+    const originalModule = jest.requireActual('../settings')
+    return {
+        ...originalModule,
+        SECRET_KEY: process.env.SECRET_KEY || 'UNSAFE_DEFAULT_SECRET_KEY',
+    }
+})
 
 describe('Chat API', () => {
     let app: Express
@@ -866,6 +886,132 @@ describe('Chat API', () => {
             expect(chatAgent.runChatAgent).toHaveBeenCalled()
             const runChatArgs = (chatAgent.runChatAgent as jest.Mock).mock.calls[0][0]
             expect(runChatArgs.llmProvider).toBe('openai') // Default LLM_PROVIDER from mocked settings
+        })
+    })
+
+    describe('Query Rewrite', () => {
+        beforeEach(() => {
+            jest.clearAllMocks()
+        })
+
+        describe('rewrite', () => {
+            it('should call LLMService.getLLM with correct parameters', async () => {
+                const mockInvoke = jest.fn().mockResolvedValue({
+                    content: 'How to authenticate users with API?',
+                })
+
+                const mockLLM = {
+                    invoke: mockInvoke,
+                }
+
+                ;(LLMService.getLLM as jest.Mock).mockReturnValue(mockLLM)
+
+                const query = 'how to auth users api'
+                const result = await rewrite(query, [])
+
+                expect(LLMService.getLLM).toHaveBeenCalledWith(0.3)
+                expect(mockInvoke).toHaveBeenCalledWith([
+                    { role: 'system', content: expect.any(String) },
+                    { role: 'user', content: expect.stringContaining(query) },
+                ])
+                expect(result).toBe('How to authenticate users with API?')
+            })
+
+            it('should include conversation history in the prompt', async () => {
+                const mockInvoke = jest.fn().mockResolvedValue({
+                    content: 'Tell me more about database migrations',
+                })
+
+                const mockLLM = {
+                    invoke: mockInvoke,
+                }
+
+                ;(LLMService.getLLM as jest.Mock).mockReturnValue(mockLLM)
+
+                const query = 'tell me more'
+                const conversationHistory = [
+                    { role: 'user' as const, content: 'What are database migrations?' },
+                    { role: 'assistant' as const, content: 'Database migrations are...' },
+                ]
+
+                await rewrite(query, conversationHistory)
+
+                const callArgs = mockInvoke.mock.calls[0][0]
+                const userMessage = callArgs[1].content
+
+                expect(userMessage).toContain('CONVERSATION CONTEXT')
+                expect(userMessage).toContain('What are database migrations?')
+                expect(userMessage).toContain('Database migrations are...')
+            })
+
+            it('should return original query on API error', async () => {
+                const mockInvoke = jest.fn().mockRejectedValue(new Error('API Error'))
+
+                const mockLLM = {
+                    invoke: mockInvoke,
+                }
+
+                ;(LLMService.getLLM as jest.Mock).mockReturnValue(mockLLM)
+
+                const query = 'how to auth'
+                const result = await rewrite(query, [])
+
+                expect(result).toBe(query) // Should fall back to original
+            })
+
+            it('should return original query when LLM returns empty response', async () => {
+                const mockInvoke = jest.fn().mockResolvedValue({
+                    content: '',
+                })
+
+                const mockLLM = {
+                    invoke: mockInvoke,
+                }
+
+                ;(LLMService.getLLM as jest.Mock).mockReturnValue(mockLLM)
+
+                const query = 'how to auth'
+                const result = await rewrite(query, [])
+
+                expect(result).toBe(query)
+            })
+
+            it('should limit conversation history to last 3 conversation pairs', async () => {
+                const mockInvoke = jest.fn().mockResolvedValue({
+                    content: 'Enhanced query',
+                })
+
+                const mockLLM = {
+                    invoke: mockInvoke,
+                }
+
+                ;(LLMService.getLLM as jest.Mock).mockReturnValue(mockLLM)
+
+                const query = 'tell me more'
+                const conversationHistory = [
+                    { role: 'user' as const, content: 'Message 1' },
+                    { role: 'assistant' as const, content: 'Response 1' },
+                    { role: 'user' as const, content: 'Message 2' },
+                    { role: 'assistant' as const, content: 'Response 2' },
+                    { role: 'user' as const, content: 'Message 3' },
+                    { role: 'assistant' as const, content: 'Response 3' },
+                    { role: 'user' as const, content: 'Message 4' },
+                    { role: 'assistant' as const, content: 'Response 4' },
+                ]
+
+                await rewrite(query, conversationHistory)
+
+                const callArgs = mockInvoke.mock.calls[0][0]
+                const userMessage = callArgs[1].content
+
+                // Should only include last 3 messages from the 8-element array
+                // slice(-3) gives indices [5, 6, 7] which are: Response 3, Message 4, Response 4
+                expect(userMessage).toContain('Response 3')
+                expect(userMessage).toContain('Message 4')
+                expect(userMessage).toContain('Response 4')
+                expect(userMessage).not.toContain('Message 1')
+                expect(userMessage).not.toContain('Message 2')
+            })
         })
     })
 })
